requirements.txt
langchain
langchain-community
pypdf
faiss-cpu
langchain-google-genai  
python-dotenv